+++ ppgan/datasets/paired_image_datasets.py
@@ -2,6 +2,7 @@ import paddle
 from paddle.vision.transforms.functional import normalize
 # from basicsr.utils.registry import DATASET_REGISTRY
 from .builder import DATASETS
+from ppgan.utils.gfpgan_tools import *
 
 
 @DATASETS.register()
@@ -119,1128 +120,3 @@ class PairedImageDataset(paddle.io.Dataset):
 
     def __len__(self):
         return len(self.paths)
-
-
-#底下是工具
-
-import cv2
-import math
-import numpy as np
-import random
-import os
-import paddle
-import paddle.nn.functional as F
-from paddle.vision.transforms.functional import normalize
-
-
-def _blend(img1, img2, ratio):
-    ratio = float(ratio)
-    bound = 1.0 if paddle.is_floating_point(img1) else 255.0
-    # print(bound)
-    return (ratio * img1 + (1.0 - ratio) * img2).clip(0, bound)
-
-
-def _get_image_num_channels(img):
-    if img.ndim == 2:
-        return 1
-    elif img.ndim > 2:
-        return img.shape[-3]
-
-    raise TypeError("Input ndim should be 2 or more. Got {}".format(img.ndim))
-
-
-def _rgb2hsv(img):
-    r, g, b = img.unbind(axis=-3)
-
-    # Implementation is based on https://github.com/python-pillow/Pillow/blob/4174d4267616897df3746d315d5a2d0f82c656ee/
-    # src/libImaging/Convert.c#L330
-    maxc = paddle.max(img, axis=-3)
-    minc = paddle.min(img, axis=-3)
-
-    # The algorithm erases S and H channel where `maxc = minc`. This avoids NaN
-    # from happening in the results, because
-    #   + S channel has division by `maxc`, which is zero only if `maxc = minc`
-    #   + H channel has division by `(maxc - minc)`.
-    #
-    # Instead of overwriting NaN afterwards, we just prevent it from occuring so
-    # we don't need to deal with it in case we save the NaN in a buffer in
-    # backprop, if it is ever supported, but it doesn't hurt to do so.
-    eqc = maxc == minc
-
-    cr = maxc - minc
-    # Since `eqc => cr = 0`, replacing denominator with 1 when `eqc` is fine.
-    ones = paddle.ones_like(maxc)
-    s = cr / paddle.where(eqc, ones, maxc)
-    # Note that `eqc => maxc = minc = r = g = b`. So the following calculation
-    # of `h` would reduce to `bc - gc + 2 + rc - bc + 4 + rc - bc = 6` so it
-    # would not matter what values `rc`, `gc`, and `bc` have here, and thus
-    # replacing denominator with 1 when `eqc` is fine.
-    cr_divisor = paddle.where(eqc, ones, cr)
-    rc = (maxc - r) / cr_divisor
-    gc = (maxc - g) / cr_divisor
-    bc = (maxc - b) / cr_divisor
-
-    t_zero = paddle.zeros_like(bc)
-    hr = paddle.where(maxc == r, (bc - gc), t_zero)
-    hg = paddle.where((maxc == g) & (maxc != r), (2.0 + rc - bc), t_zero)
-    hb = paddle.where((maxc != g) & (maxc != r), (4.0 + gc - rc), t_zero)
-
-    h = (hr + hg + hb)
-    h = paddle.mod((h / 6.0 + 1.0), paddle.to_tensor([1.0]))
-    return paddle.stack((h, s, maxc), axis=-3)
-
-
-def _hsv2rgb(img):
-    h, s, v = img.unbind(axis=-3)
-    i = paddle.floor(h * 6.0)
-    f = (h * 6.0) - i
-    i = paddle.cast(i, dtype='int32')
-
-    p = paddle.clip((v * (1.0 - s)), 0.0, 1.0)
-    q = paddle.clip((v * (1.0 - s * f)), 0.0, 1.0)
-    t = paddle.clip((v * (1.0 - s * (1.0 - f))), 0.0, 1.0)
-    i = i % 6
-
-    mask = i.unsqueeze(axis=-3) == paddle.arange(6).reshape([-1, 1, 1])
-
-    a1 = paddle.stack((v, q, p, p, t, v), axis=-3)
-    a2 = paddle.stack((t, v, v, q, p, p), axis=-3)
-    a3 = paddle.stack((p, p, t, v, v, q), axis=-3)
-    a4 = paddle.stack((a1, a2, a3), axis=-4)
-    t_zero = paddle.zeros_like(mask, dtype='float32')
-    t_ones = paddle.ones_like(mask, dtype='float32')
-    mask = paddle.where(mask, t_ones, t_zero)
-    return paddle.einsum("...ijk, ...xijk -> ...xjk", mask, a4)
-
-
-def rgb_to_grayscale(img, num_output_channels=1):
-    if img.ndim < 3:
-        raise TypeError(
-            "Input image tensor should have at least 3 axisensions, but found {}"
-            .format(img.ndim))
-
-    if num_output_channels not in (1, 3):
-        raise ValueError('num_output_channels should be either 1 or 3')
-
-    r, g, b = img.unbind(axis=-3)
-    l_img = (0.2989 * r + 0.587 * g + 0.114 * b)
-    l_img = l_img.unsqueeze(axis=-3)
-
-    if num_output_channels == 3:
-        return l_img.expand(img.shape)
-
-    return l_img
-
-
-def adjust_brightness(img, brightness_factor):
-    if brightness_factor < 0:
-        raise ValueError('brightness_factor ({}) is not non-negative.'.format(
-            brightness_factor))
-
-    return _blend(img, paddle.zeros_like(img), brightness_factor)
-
-
-def adjust_contrast(img, contrast_factor):
-    if contrast_factor < 0:
-        raise ValueError(
-            'contrast_factor ({}) is not non-negative.'.format(contrast_factor))
-
-    dtype = img.dtype if paddle.is_floating_point(img) else paddle.float32
-    mean = paddle.mean(paddle.cast(rgb_to_grayscale(img), dtype=dtype),
-                       axis=(-3, -2, -1),
-                       keepdim=True)
-
-    return _blend(img, mean, contrast_factor)
-
-
-def adjust_hue(img, hue_factor):
-    if not (-0.5 <= hue_factor <= 0.5):
-        raise ValueError(
-            'hue_factor ({}) is not in [-0.5, 0.5].'.format(hue_factor))
-
-    if not (isinstance(img, paddle.Tensor)):
-        raise TypeError('Input img should be Tensor image')
-
-    if _get_image_num_channels(img) == 1:  # Match PIL behaviour
-        return img
-
-    orig_dtype = img.dtype
-    if img.dtype == paddle.uint8:
-        img = paddle.cast(img, dtype='float32') / 255.0
-
-    img = _rgb2hsv(img)
-    h, s, v = img.unbind(axis=-3)
-    h = (h + hue_factor) % 1.0
-    img = paddle.stack((h, s, v), axis=-3)
-    img_hue_adj = _hsv2rgb(img)
-
-    if orig_dtype == paddle.uint8:
-        img_hue_adj = paddle.cast(img_hue_adj * 255.0, dtype=orig_dtype)
-
-    return img_hue_adj
-
-
-def adjust_saturation(img, saturation_factor):
-    if saturation_factor < 0:
-        raise ValueError('saturation_factor ({}) is not non-negative.'.format(
-            saturation_factor))
-
-    return _blend(img, rgb_to_grayscale(img), saturation_factor)
-
-
-def generate_gaussian_noise(img, sigma=10, gray_noise=False):
-    """Generate Gaussian noise.
-
-    Args:
-        img (Numpy array): Input image, shape (h, w, c), range [0, 1], float32.
-        sigma (float): Noise scale (measured in range 255). Default: 10.
-
-    Returns:
-        (Numpy array): Returned noisy image, shape (h, w, c), range[0, 1],
-            float32.
-    """
-    if gray_noise:
-        noise = np.float32(np.random.randn(*img.shape[0:2])) * sigma / 255.0
-        noise = np.expand_dims(noise, axis=2).repeat(3, axis=2)
-    else:
-        noise = np.float32(np.random.randn(*img.shape)) * sigma / 255.0
-    return noise
-
-
-def random_generate_gaussian_noise(img, sigma_range=(0, 10), gray_prob=0):
-    sigma = np.random.uniform(sigma_range[0], sigma_range[1])
-    if np.random.uniform() < gray_prob:
-        gray_noise = True
-    else:
-        gray_noise = False
-    return generate_gaussian_noise(img, sigma, gray_noise)
-
-
-def random_add_gaussian_noise(img, sigma_range=(0, 1.0), gray_prob=0, clip=\
-    True, rounds=False):
-    noise = random_generate_gaussian_noise(img, sigma_range, gray_prob)
-    out = img + noise
-    if clip and rounds:
-        out = np.clip((out * 255.0).round(), 0, 255) / 255.0
-    elif clip:
-        out = np.clip(out, 0, 1)
-    elif rounds:
-        out = (out * 255.0).round() / 255.0
-    return out
-
-
-def add_jpg_compression(img, quality=90):
-    """Add JPG compression artifacts.
-
-    Args:
-        img (Numpy array): Input image, shape (h, w, c), range [0, 1], float32.
-        quality (float): JPG compression quality. 0 for lowest quality, 100 for
-            best quality. Default: 90.
-
-    Returns:
-        (Numpy array): Returned image after JPG, shape (h, w, c), range[0, 1],
-            float32.
-    """
-    img = np.clip(img, 0, 1)
-    encode_param = [int(cv2.IMWRITE_JPEG_QUALITY), quality]
-    _, encimg = cv2.imencode('.jpg', img * 255.0, encode_param)
-    img = np.float32(cv2.imdecode(encimg, 1)) / 255.0
-    return img
-
-
-def random_add_jpg_compression(img, quality_range=(90, 100)):
-    """Randomly add JPG compression artifacts.
-
-    Args:
-        img (Numpy array): Input image, shape (h, w, c), range [0, 1], float32.
-        quality_range (tuple[float] | list[float]): JPG compression quality
-            range. 0 for lowest quality, 100 for best quality.
-            Default: (90, 100).
-
-    Returns:
-        (Numpy array): Returned image after JPG, shape (h, w, c), range[0, 1],
-            float32.
-    """
-    quality = int(np.random.uniform(quality_range[0], quality_range[1]))
-    return add_jpg_compression(img, quality)
-
-
-def random_mixed_kernels(kernel_list,
-                         kernel_prob,
-                         kernel_size=21,
-                         sigma_x_range=(0.6, 5),
-                         sigma_y_range=(0.6, 5),
-                         rotation_range=(-math.pi, math.pi),
-                         betag_range=(0.5, 8),
-                         betap_range=(0.5, 8),
-                         noise_range=None):
-    """Randomly generate mixed kernels.
-
-    Args:
-        kernel_list (tuple): a list name of kernel types,
-            support ['iso', 'aniso', 'skew', 'generalized', 'plateau_iso',
-            'plateau_aniso']
-        kernel_prob (tuple): corresponding kernel probability for each
-            kernel type
-        kernel_size (int):
-        sigma_x_range (tuple): [0.6, 5]
-        sigma_y_range (tuple): [0.6, 5]
-        rotation range (tuple): [-math.pi, math.pi]
-        beta_range (tuple): [0.5, 8]
-        noise_range(tuple, optional): multiplicative kernel noise,
-            [0.75, 1.25]. Default: None
-
-    Returns:
-        kernel (ndarray):
-    """
-    kernel_type = random.choices(kernel_list, kernel_prob)[0]
-    if kernel_type == 'iso':
-        kernel = random_bivariate_Gaussian(kernel_size,
-                                           sigma_x_range,
-                                           sigma_y_range,
-                                           rotation_range,
-                                           noise_range=noise_range,
-                                           isotropic=True)
-    elif kernel_type == 'aniso':
-        kernel = random_bivariate_Gaussian(kernel_size,
-                                           sigma_x_range,
-                                           sigma_y_range,
-                                           rotation_range,
-                                           noise_range=noise_range,
-                                           isotropic=False)
-    return kernel
-
-
-def random_bivariate_Gaussian(kernel_size,
-                              sigma_x_range,
-                              sigma_y_range,
-                              rotation_range,
-                              noise_range=None,
-                              isotropic=True):
-    """Randomly generate bivariate isotropic or anisotropic Gaussian kernels.
-
-    In the isotropic mode, only `sigma_x_range` is used. `sigma_y_range` and `rotation_range` is ignored.
-
-    Args:
-        kernel_size (int):
-        sigma_x_range (tuple): [0.6, 5]
-        sigma_y_range (tuple): [0.6, 5]
-        rotation range (tuple): [-math.pi, math.pi]
-        noise_range(tuple, optional): multiplicative kernel noise,
-            [0.75, 1.25]. Default: None
-
-    Returns:
-        kernel (ndarray):
-    """
-    assert kernel_size % 2 == 1, 'Kernel size must be an odd number.'
-    assert sigma_x_range[0] < sigma_x_range[1], 'Wrong sigma_x_range.'
-    sigma_x = np.random.uniform(sigma_x_range[0], sigma_x_range[1])
-    if isotropic is False:
-        assert sigma_y_range[0] < sigma_y_range[1], 'Wrong sigma_y_range.'
-        assert rotation_range[0] < rotation_range[1], 'Wrong rotation_range.'
-        sigma_y = np.random.uniform(sigma_y_range[0], sigma_y_range[1])
-        rotation = np.random.uniform(rotation_range[0], rotation_range[1])
-    else:
-        sigma_y = sigma_x
-        rotation = 0
-    kernel = bivariate_Gaussian(kernel_size,
-                                sigma_x,
-                                sigma_y,
-                                rotation,
-                                isotropic=isotropic)
-    if noise_range is not None:
-        assert noise_range[0] < noise_range[1], 'Wrong noise range.'
-        noise = np.random.uniform(noise_range[0], noise_range[1], size=\
-            kernel.shape)
-        kernel = kernel * noise
-    kernel = kernel / np.sum(kernel)
-    return kernel
-
-
-def bivariate_Gaussian(kernel_size,
-                       sig_x,
-                       sig_y,
-                       theta,
-                       grid=None,
-                       isotropic=True):
-    """Generate a bivariate isotropic or anisotropic Gaussian kernel.
-
-    In the isotropic mode, only `sig_x` is used. `sig_y` and `theta` is ignored.
-
-    Args:
-        kernel_size (int):
-        sig_x (float):
-        sig_y (float):
-        theta (float): Radian measurement.
-        grid (ndarray, optional): generated by :func:`mesh_grid`,
-            with the shape (K, K, 2), K is the kernel size. Default: None
-        isotropic (bool):
-
-    Returns:
-        kernel (ndarray): normalized kernel.
-    """
-    if grid is None:
-        grid, _, _ = mesh_grid(kernel_size)
-    if isotropic:
-        sigma_matrix = np.array([[sig_x**2, 0], [0, sig_x**2]])
-    else:
-        sigma_matrix = sigma_matrix2(sig_x, sig_y, theta)
-    kernel = pdf2(sigma_matrix, grid)
-    kernel = kernel / np.sum(kernel)
-    return kernel
-
-
-def sigma_matrix2(sig_x, sig_y, theta):
-    """Calculate the rotated sigma matrix (two dimensional matrix).
-
-    Args:
-        sig_x (float):
-        sig_y (float):
-        theta (float): Radian measurement.
-
-    Returns:
-        ndarray: Rotated sigma matrix.
-    """
-    d_matrix = np.array([[sig_x**2, 0], [0, sig_y**2]])
-    u_matrix = np.array([[np.cos(theta), -np.sin(theta)],
-                         [np.sin(theta), np.cos(theta)]])
-    return np.dot(u_matrix, np.dot(d_matrix, u_matrix.T))
-
-
-def mesh_grid(kernel_size):
-    """Generate the mesh grid, centering at zero.
-
-    Args:
-        kernel_size (int):
-
-    Returns:
-        xy (ndarray): with the shape (kernel_size, kernel_size, 2)
-        xx (ndarray): with the shape (kernel_size, kernel_size)
-        yy (ndarray): with the shape (kernel_size, kernel_size)
-    """
-    ax = np.arange(-kernel_size // 2 + 1.0, kernel_size // 2 + 1.0)
-    xx, yy = np.meshgrid(ax, ax)
-    xy = np.hstack((xx.reshape((kernel_size * kernel_size, 1)),
-                    yy.reshape(kernel_size * kernel_size,
-                               1))).reshape(kernel_size, kernel_size, 2)
-    return xy, xx, yy
-
-
-def pdf2(sigma_matrix, grid):
-    """Calculate PDF of the bivariate Gaussian distribution.
-
-    Args:
-        sigma_matrix (ndarray): with the shape (2, 2)
-        grid (ndarray): generated by :func:`mesh_grid`,
-            with the shape (K, K, 2), K is the kernel size.
-
-    Returns:
-        kernel (ndarrray): un-normalized kernel.
-    """
-    inverse_sigma = np.linalg.inv(sigma_matrix)
-    kernel = np.exp(-0.5 * np.sum(np.dot(grid, inverse_sigma) * grid, 2))
-    return kernel
-
-
-import os.path as osp
-
-
-def paths_from_folder(folder):
-    """Generate paths from folder.
-
-    Args:
-        folder (str): Folder path.
-
-    Returns:
-        list[str]: Returned path list.
-    """
-    paths = list(scandir(folder))
-    paths = [osp.join(folder, path) for path in paths]
-    return paths
-
-
-def scandir(dir_path, suffix=None, recursive=False, full_path=False):
-    """Scan a directory to find the interested files.
-
-    Args:
-        dir_path (str): Path of the directory.
-        suffix (str | tuple(str), optional): File suffix that we are
-            interested in. Default: None.
-        recursive (bool, optional): If set to True, recursively scan the
-            directory. Default: False.
-        full_path (bool, optional): If set to True, include the dir_path.
-            Default: False.
-
-    Returns:
-        A generator for all the interested files with relative paths.
-    """
-    if suffix is not None and not isinstance(suffix, (str, tuple)):
-        raise TypeError('"suffix" must be a string or tuple of strings')
-    root = dir_path
-
-    def _scandir(dir_path, suffix, recursive):
-        for entry in os.scandir(dir_path):
-            if not entry.name.startswith('.') and entry.is_file():
-                if full_path:
-                    return_path = entry.path
-                else:
-                    return_path = osp.relpath(entry.path, root)
-                if suffix is None:
-                    yield return_path
-                elif return_path.endswith(suffix):
-                    yield return_path
-            elif recursive:
-                yield from _scandir(entry.path, suffix=suffix, recursive=\
-                    recursive)
-            else:
-                continue
-
-    return _scandir(dir_path, suffix=suffix, recursive=recursive)
-
-
-from abc import ABCMeta
-from abc import abstractmethod
-
-
-class BaseStorageBackend(metaclass=ABCMeta):
-    """Abstract class of storage backends.
-
-    All backends need to implement two apis: ``get()`` and ``get_text()``.
-    ``get()`` reads the file as a byte stream and ``get_text()`` reads the file
-    as texts.
-    """
-    @abstractmethod
-    def get(self, filepath):
-        pass
-
-    @abstractmethod
-    def get_text(self, filepath):
-        pass
-
-
-class MemcachedBackend(BaseStorageBackend):
-    """Memcached storage backend.
-
-    Attributes:
-        server_list_cfg (str): Config file for memcached server list.
-        client_cfg (str): Config file for memcached client.
-        sys_path (str | None): Additional path to be appended to `sys.path`.
-            Default: None.
-    """
-    def __init__(self, server_list_cfg, client_cfg, sys_path=None):
-        if sys_path is not None:
-            import sys
-            sys.path.append(sys_path)
-        try:
-            import mc
-        except ImportError:
-            raise ImportError(
-                'Please install memcached to enable MemcachedBackend.')
-        self.server_list_cfg = server_list_cfg
-        self.client_cfg = client_cfg
-        self._client = mc.MemcachedClient.GetInstance(self.server_list_cfg,
-                                                      self.client_cfg)
-        self._mc_buffer = mc.pyvector()
-
-    def get(self, filepath):
-        filepath = str(filepath)
-        import mc
-        self._client.Get(filepath, self._mc_buffer)
-        value_buf = mc.ConvertBuffer(self._mc_buffer)
-        return value_buf
-
-    def get_text(self, filepath):
-        raise NotImplementedError
-
-
-class HardDiskBackend(BaseStorageBackend):
-    """Raw hard disks storage backend."""
-    def get(self, filepath):
-        filepath = str(filepath)
-        with open(filepath, 'rb') as f:
-            value_buf = f.read()
-        return value_buf
-
-    def get_text(self, filepath):
-        filepath = str(filepath)
-        with open(filepath, 'r') as f:
-            value_buf = f.read()
-        return value_buf
-
-
-class LmdbBackend(BaseStorageBackend):
-    """Lmdb storage backend.
-
-    Args:
-        db_paths (str | list[str]): Lmdb database paths.
-        client_keys (str | list[str]): Lmdb client keys. Default: 'default'.
-        readonly (bool, optional): Lmdb environment parameter. If True,
-            disallow any write operations. Default: True.
-        lock (bool, optional): Lmdb environment parameter. If False, when
-            concurrent access occurs, do not lock the database. Default: False.
-        readahead (bool, optional): Lmdb environment parameter. If False,
-            disable the OS filesystem readahead mechanism, which may improve
-            random read performance when a database is larger than RAM.
-            Default: False.
-
-    Attributes:
-        db_paths (list): Lmdb database path.
-        _client (list): A list of several lmdb envs.
-    """
-    def __init__(self,
-                 db_paths,
-                 client_keys='default',
-                 readonly=True,
-                 lock=False,
-                 readahead=False,
-                 **kwargs):
-        try:
-            import lmdb
-        except ImportError:
-            raise ImportError('Please install lmdb to enable LmdbBackend.')
-        if isinstance(client_keys, str):
-            client_keys = [client_keys]
-        if isinstance(db_paths, list):
-            self.db_paths = [str(v) for v in db_paths]
-        elif isinstance(db_paths, str):
-            self.db_paths = [str(db_paths)]
-        assert len(client_keys) == len(
-            self.db_paths
-        ), f'client_keys and db_paths should have the same length, but received {len(client_keys)} and {len(self.db_paths)}.'
-        self._client = {}
-        for client, path in zip(client_keys, self.db_paths):
-            self._client[client] = lmdb.open(path, readonly=readonly, lock=\
-                lock, readahead=readahead, **kwargs)
-
-    def get(self, filepath, client_key):
-        """Get values according to the filepath from one lmdb named client_key.
-
-        Args:
-            filepath (str | obj:`Path`): Here, filepath is the lmdb key.
-            client_key (str): Used for distinguishing different lmdb envs.
-        """
-        filepath = str(filepath)
-        assert client_key in self._client, f'client_key {client_key} is not in lmdb clients.'
-        client = self._client[client_key]
-        with client.begin(write=False) as txn:
-            value_buf = txn.get(filepath.encode('ascii'))
-        return value_buf
-
-    def get_text(self, filepath):
-        raise NotImplementedError
-
-
-class FileClient(object):
-    """A general file client to access files in different backend.
-
-    The client loads a file or text in a specified backend from its path
-    and return it as a binary file. it can also register other backend
-    accessor with a given name and backend class.
-
-    Attributes:
-        backend (str): The storage backend type. Options are "disk",
-            "memcached" and "lmdb".
-        client (:obj:`BaseStorageBackend`): The backend object.
-    """
-    _backends = {
-        'disk': HardDiskBackend,
-        'memcached': MemcachedBackend,
-        'lmdb': LmdbBackend
-    }
-
-    def __init__(self, backend='disk', **kwargs):
-        if backend not in self._backends:
-            raise ValueError(
-                f'Backend {backend} is not supported. Currently supported ones are {list(self._backends.keys())}'
-            )
-        self.backend = backend
-        self.client = self._backends[backend](**kwargs)
-
-    def get(self, filepath, client_key='default'):
-        if self.backend == 'lmdb':
-            return self.client.get(filepath, client_key)
-        else:
-            return self.client.get(filepath)
-
-    def get_text(self, filepath):
-        return self.client.get_text(filepath)
-
-
-def imfrombytes(content, flag='color', float32=False):
-    """Read an image from bytes.
-
-    Args:
-        content (bytes): Image bytes got from files or other streams.
-        flag (str): Flags specifying the color type of a loaded image,
-            candidates are `color`, `grayscale` and `unchanged`.
-        float32 (bool): Whether to change to float32., If True, will also norm
-            to [0, 1]. Default: False.
-
-    Returns:
-        ndarray: Loaded image array.
-    """
-    img_np = np.frombuffer(content, np.uint8)
-    imread_flags = {
-        'color': cv2.IMREAD_COLOR,
-        'grayscale': cv2.IMREAD_GRAYSCALE,
-        'unchanged': cv2.IMREAD_UNCHANGED
-    }
-    img = cv2.imdecode(img_np, imread_flags[flag])
-    if float32:
-        img = img.astype(np.float32) / 255.0
-    return img
-
-
-def img2tensor(imgs, bgr2rgb=True, float32=True):
-    """Numpy array to tensor.
-
-    Args:
-        imgs (list[ndarray] | ndarray): Input images.
-        bgr2rgb (bool): Whether to change bgr to rgb.
-        float32 (bool): Whether to change to float32.
-
-    Returns:
-        list[tensor] | tensor: Tensor images. If returned results only have
-            one element, just return tensor.
-    """
-    def _totensor(img, bgr2rgb, float32):
-        if img.shape[2] == 3 and bgr2rgb:
-            if img.dtype == 'float64':
-                img = img.astype('float32')
-            img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-
-        # img = paddle.to_tensor(img.transpose(2, 0, 1))
-
-        return img.transpose(2, 0, 1)
-
-    if isinstance(imgs, list):
-        return [_totensor(img, bgr2rgb, float32) for img in imgs]
-    else:
-        return _totensor(imgs, bgr2rgb, float32)
-
-
-def mod_crop(img, scale):
-    """Mod crop images, used during testing.
-
-    Args:
-        img (ndarray): Input image.
-        scale (int): Scale factor.
-
-    Returns:
-        ndarray: Result image.
-    """
-    img = img.copy()
-    if img.ndim in (2, 3):
-        h, w = img.shape[0], img.shape[1]
-        h_remainder, w_remainder = h % scale, w % scale
-        img = img[:h - h_remainder, :w - w_remainder, ...]
-    else:
-        raise ValueError(f'Wrong img ndim: {img.ndim}.')
-    return img
-
-
-def paired_random_crop(img_gts, img_lqs, gt_patch_size, scale, gt_path=None):
-    """Paired random crop. Support Numpy array and Tensor inputs.
-
-    It crops lists of lq and gt images with corresponding locations.
-
-    Args:
-        img_gts (list[ndarray] | ndarray | list[Tensor] | Tensor): GT images. Note that all images
-            should have the same shape. If the input is an ndarray, it will
-            be transformed to a list containing itself.
-        img_lqs (list[ndarray] | ndarray): LQ images. Note that all images
-            should have the same shape. If the input is an ndarray, it will
-            be transformed to a list containing itself.
-        gt_patch_size (int): GT patch size.
-        scale (int): Scale factor.
-        gt_path (str): Path to ground-truth. Default: None.
-
-    Returns:
-        list[ndarray] | ndarray: GT images and LQ images. If returned results
-            only have one element, just return ndarray.
-    """
-    if not isinstance(img_gts, list):
-        img_gts = [img_gts]
-    if not isinstance(img_lqs, list):
-        img_lqs = [img_lqs]
-    input_type = 'Tensor' if isinstance(img_gts[0], paddle.Tensor) else 'Numpy'
-    if input_type == 'Tensor':
-        h_lq, w_lq = img_lqs[0].size()[-2:]
-        h_gt, w_gt = img_gts[0].size()[-2:]
-    else:
-        h_lq, w_lq = img_lqs[0].shape[0:2]
-        h_gt, w_gt = img_gts[0].shape[0:2]
-    lq_patch_size = gt_patch_size // scale
-    if h_gt != h_lq * scale or w_gt != w_lq * scale:
-        raise ValueError(
-            f'Scale mismatches. GT ({h_gt}, {w_gt}) is not {scale}x ',
-            f'multiplication of LQ ({h_lq}, {w_lq}).')
-    if h_lq < lq_patch_size or w_lq < lq_patch_size:
-        raise ValueError(
-            f'LQ ({h_lq}, {w_lq}) is smaller than patch size ({lq_patch_size}, {lq_patch_size}). Please remove {gt_path}.'
-        )
-    top = random.randint(0, h_lq - lq_patch_size)
-    left = random.randint(0, w_lq - lq_patch_size)
-    if input_type == 'Tensor':
-        img_lqs = [
-            v[:, :, top:top + lq_patch_size, left:left + lq_patch_size]
-            for v in img_lqs
-        ]
-    else:
-        img_lqs = [
-            v[top:top + lq_patch_size, left:left + lq_patch_size, ...]
-            for v in img_lqs
-        ]
-    top_gt, left_gt = int(top * scale), int(left * scale)
-    if input_type == 'Tensor':
-        img_gts = [
-            v[:, :, top_gt:top_gt + gt_patch_size,
-              left_gt:left_gt + gt_patch_size] for v in img_gts
-        ]
-    else:
-        img_gts = [
-            v[top_gt:top_gt + gt_patch_size, left_gt:left_gt + gt_patch_size,
-              ...] for v in img_gts
-        ]
-    if len(img_gts) == 1:
-        img_gts = img_gts[0]
-    if len(img_lqs) == 1:
-        img_lqs = img_lqs[0]
-    return img_gts, img_lqs
-
-
-def augment(imgs, hflip=True, rotation=True, flows=None, return_status=False):
-    """Augment: horizontal flips OR rotate (0, 90, 180, 270 degrees).
-
-    We use vertical flip and transpose for rotation implementation.
-    All the images in the list use the same augmentation.
-
-    Args:
-        imgs (list[ndarray] | ndarray): Images to be augmented. If the input
-            is an ndarray, it will be transformed to a list.
-        hflip (bool): Horizontal flip. Default: True.
-        rotation (bool): Ratotation. Default: True.
-        flows (list[ndarray]: Flows to be augmented. If the input is an
-            ndarray, it will be transformed to a list.
-            Dimension is (h, w, 2). Default: None.
-        return_status (bool): Return the status of flip and rotation.
-            Default: False.
-
-    Returns:
-        list[ndarray] | ndarray: Augmented images and flows. If returned
-            results only have one element, just return ndarray.
-
-    """
-    hflip = hflip and random.random() < 0.5
-    vflip = rotation and random.random() < 0.5
-    rot90 = rotation and random.random() < 0.5
-
-    def _augment(img):
-        if hflip:
-            cv2.flip(img, 1, img)
-        if vflip:
-            cv2.flip(img, 0, img)
-        if rot90:
-            img = img.transpose(1, 0, 2)
-        return img
-
-    def _augment_flow(flow):
-        if hflip:
-            cv2.flip(flow, 1, flow)
-            flow[:, :, 0] *= -1
-        if vflip:
-            cv2.flip(flow, 0, flow)
-            flow[:, :, 1] *= -1
-        if rot90:
-            flow = flow.transpose(1, 0, 2)
-            flow = flow[:, :, [1, 0]]
-        return flow
-
-    if not isinstance(imgs, list):
-        imgs = [imgs]
-    imgs = [_augment(img) for img in imgs]
-    if len(imgs) == 1:
-        imgs = imgs[0]
-    if flows is not None:
-        if not isinstance(flows, list):
-            flows = [flows]
-        flows = [_augment_flow(flow) for flow in flows]
-        if len(flows) == 1:
-            flows = flows[0]
-        return imgs, flows
-    elif return_status:
-        return imgs, (hflip, vflip, rot90)
-    else:
-        return imgs
-
-
-def img_rotate(img, angle, center=None, scale=1.0):
-    """Rotate image.
-
-    Args:
-        img (ndarray): Image to be rotated.
-        angle (float): Rotation angle in degrees. Positive values mean
-            counter-clockwise rotation.
-        center (tuple[int]): Rotation center. If the center is None,
-            initialize it as the center of the image. Default: None.
-        scale (float): Isotropic scale factor. Default: 1.0.
-    """
-    h, w = img.shape[:2]
-    if center is None:
-        center = w // 2, h // 2
-    matrix = cv2.getRotationMatrix2D(center, angle, scale)
-    rotated_img = cv2.warpAffine(img, matrix, (w, h))
-    return rotated_img
-
-
-def _convert_input_type_range(img):
-    """Convert the type and range of the input image.
-
-    It converts the input image to np.float32 type and range of [0, 1].
-    It is mainly used for pre-processing the input image in colorspace
-    conversion functions such as rgb2ycbcr and ycbcr2rgb.
-
-    Args:
-        img (ndarray): The input image. It accepts:
-            1. np.uint8 type with range [0, 255];
-            2. np.float32 type with range [0, 1].
-
-    Returns:
-        (ndarray): The converted image with type of np.float32 and range of
-            [0, 1].
-    """
-    img_type = img.dtype
-    img = img.astype(np.float32)
-    if img_type == np.float32:
-        pass
-    elif img_type == np.uint8:
-        img /= 255.
-    else:
-        raise TypeError(
-            f'The img type should be np.float32 or np.uint8, but got {img_type}'
-        )
-    return img
-
-
-def _convert_output_type_range(img, dst_type):
-    """Convert the type and range of the image according to dst_type.
-
-    It converts the image to desired type and range. If `dst_type` is np.uint8,
-    images will be converted to np.uint8 type with range [0, 255]. If
-    `dst_type` is np.float32, it converts the image to np.float32 type with
-    range [0, 1].
-    It is mainly used for post-processing images in colorspace conversion
-    functions such as rgb2ycbcr and ycbcr2rgb.
-
-    Args:
-        img (ndarray): The image to be converted with np.float32 type and
-            range [0, 255].
-        dst_type (np.uint8 | np.float32): If dst_type is np.uint8, it
-            converts the image to np.uint8 type with range [0, 255]. If
-            dst_type is np.float32, it converts the image to np.float32 type
-            with range [0, 1].
-
-    Returns:
-        (ndarray): The converted image with desired type and range.
-    """
-    if dst_type not in (np.uint8, np.float32):
-        raise TypeError(
-            f'The dst_type should be np.float32 or np.uint8, but got {dst_type}'
-        )
-    if dst_type == np.uint8:
-        img = img.round()
-    else:
-        img /= 255.
-    return img.astype(dst_type)
-
-
-def bgr2ycbcr(img, y_only=False):
-    """Convert a BGR image to YCbCr image.
-
-    The bgr version of rgb2ycbcr.
-    It implements the ITU-R BT.601 conversion for standard-definition
-    television. See more details in
-    https://en.wikipedia.org/wiki/YCbCr#ITU-R_BT.601_conversion.
-
-    It differs from a similar function in cv2.cvtColor: `BGR <-> YCrCb`.
-    In OpenCV, it implements a JPEG conversion. See more details in
-    https://en.wikipedia.org/wiki/YCbCr#JPEG_conversion.
-
-    Args:
-        img (ndarray): The input image. It accepts:
-            1. np.uint8 type with range [0, 255];
-            2. np.float32 type with range [0, 1].
-        y_only (bool): Whether to only return Y channel. Default: False.
-
-    Returns:
-        ndarray: The converted YCbCr image. The output image has the same type
-            and range as input image.
-    """
-    img_type = img.dtype
-    img = _convert_input_type_range(img)
-    if y_only:
-        out_img = np.dot(img, [24.966, 128.553, 65.481]) + 16.0
-    else:
-        out_img = np.matmul(
-            img, [[24.966, 112.0, -18.214], [128.553, -74.203, -93.786],
-                  [65.481, -37.797, 112.0]]) + [16, 128, 128]
-    out_img = _convert_output_type_range(out_img, img_type)
-    return out_img
-
-
-def paired_paths_from_folder(folders, keys, filename_tmpl):
-    """Generate paired paths from folders.
-
-    Args:
-        folders (list[str]): A list of folder path. The order of list should
-            be [input_folder, gt_folder].
-        keys (list[str]): A list of keys identifying folders. The order should
-            be in consistent with folders, e.g., ['lq', 'gt'].
-        filename_tmpl (str): Template for each filename. Note that the
-            template excludes the file extension. Usually the filename_tmpl is
-            for files in the input folder.
-
-    Returns:
-        list[str]: Returned path list.
-    """
-    assert len(folders) == 2, (
-        'The len of folders should be 2 with [input_folder, gt_folder]. '
-        f'But got {len(folders)}')
-    assert len(
-        keys
-    ) == 2, f'The len of keys should be 2 with [input_key, gt_key]. But got {len(keys)}'
-    input_folder, gt_folder = folders
-    input_key, gt_key = keys
-
-    input_paths = list(scandir(input_folder))
-    gt_paths = list(scandir(gt_folder))
-    assert len(input_paths) == len(gt_paths), (
-        f'{input_key} and {gt_key} datasets have different number of images: '
-        f'{len(input_paths)}, {len(gt_paths)}.')
-    paths = []
-    for gt_path in gt_paths:
-        basename, ext = osp.splitext(osp.basename(gt_path))
-        input_name = f'{filename_tmpl.format(basename)}{ext}'
-        input_path = osp.join(input_folder, input_name)
-        assert input_name in input_paths, f'{input_name} is not in {input_key}_paths.'
-        gt_path = osp.join(gt_folder, gt_path)
-        paths.append(
-            dict([(f'{input_key}_path', input_path),
-                  (f'{gt_key}_path', gt_path)]))
-    return paths
-
-
-def paired_paths_from_lmdb(folders, keys):
-    """Generate paired paths from lmdb files.
-
-    Contents of lmdb. Taking the `lq.lmdb` for example, the file structure is:
-
-    lq.lmdb
-    ├── data.mdb
-    ├── lock.mdb
-    ├── meta_info.txt
-
-    The data.mdb and lock.mdb are standard lmdb files and you can refer to
-    https://lmdb.readthedocs.io/en/release/ for more details.
-
-    The meta_info.txt is a specified txt file to record the meta information
-    of our datasets. It will be automatically created when preparing
-    datasets by our provided dataset tools.
-    Each line in the txt file records
-    1)image name (with extension),
-    2)image shape,
-    3)compression level, separated by a white space.
-    Example: `baboon.png (120,125,3) 1`
-
-    We use the image name without extension as the lmdb key.
-    Note that we use the same key for the corresponding lq and gt images.
-
-    Args:
-        folders (list[str]): A list of folder path. The order of list should
-            be [input_folder, gt_folder].
-        keys (list[str]): A list of keys identifying folders. The order should
-            be in consistent with folders, e.g., ['lq', 'gt'].
-            Note that this key is different from lmdb keys.
-
-    Returns:
-        list[str]: Returned path list.
-    """
-    assert len(folders) == 2, (
-        'The len of folders should be 2 with [input_folder, gt_folder]. '
-        f'But got {len(folders)}')
-    assert len(
-        keys
-    ) == 2, f'The len of keys should be 2 with [input_key, gt_key]. But got {len(keys)}'
-    input_folder, gt_folder = folders
-    input_key, gt_key = keys
-
-    if not (input_folder.endswith('.lmdb') and gt_folder.endswith('.lmdb')):
-        raise ValueError(
-            f'{input_key} folder and {gt_key} folder should both in lmdb '
-            f'formats. But received {input_key}: {input_folder}; '
-            f'{gt_key}: {gt_folder}')
-    # ensure that the two meta_info files are the same
-    with open(osp.join(input_folder, 'meta_info.txt')) as fin:
-        input_lmdb_keys = [line.split('.')[0] for line in fin]
-    with open(osp.join(gt_folder, 'meta_info.txt')) as fin:
-        gt_lmdb_keys = [line.split('.')[0] for line in fin]
-    if set(input_lmdb_keys) != set(gt_lmdb_keys):
-        raise ValueError(
-            f'Keys in {input_key}_folder and {gt_key}_folder are different.')
-    else:
-        paths = []
-        for lmdb_key in sorted(input_lmdb_keys):
-            paths.append(
-                dict([(f'{input_key}_path', lmdb_key),
-                      (f'{gt_key}_path', lmdb_key)]))
-        return paths
-
-
-def paired_paths_from_meta_info_file(folders, keys, meta_info_file,
-                                     filename_tmpl):
-    """Generate paired paths from an meta information file.
-
-    Each line in the meta information file contains the image names and
-    image shape (usually for gt), separated by a white space.
-
-    Example of an meta information file:
-    ```
-    0001_s001.png (480,480,3)
-    0001_s002.png (480,480,3)
-    ```
-
-    Args:
-        folders (list[str]): A list of folder path. The order of list should
-            be [input_folder, gt_folder].
-        keys (list[str]): A list of keys identifying folders. The order should
-            be in consistent with folders, e.g., ['lq', 'gt'].
-        meta_info_file (str): Path to the meta information file.
-        filename_tmpl (str): Template for each filename. Note that the
-            template excludes the file extension. Usually the filename_tmpl is
-            for files in the input folder.
-
-    Returns:
-        list[str]: Returned path list.
-    """
-    assert len(folders) == 2, (
-        'The len of folders should be 2 with [input_folder, gt_folder]. '
-        f'But got {len(folders)}')
-    assert len(
-        keys
-    ) == 2, f'The len of keys should be 2 with [input_key, gt_key]. But got {len(keys)}'
-    input_folder, gt_folder = folders
-    input_key, gt_key = keys
-
-    with open(meta_info_file, 'r') as fin:
-        gt_names = [line.strip().split(' ')[0] for line in fin]
-
-    paths = []
-    for gt_name in gt_names:
-        basename, ext = osp.splitext(osp.basename(gt_name))
-        input_name = f'{filename_tmpl.format(basename)}{ext}'
-        input_path = osp.join(input_folder, input_name)
-        gt_path = osp.join(gt_folder, gt_name)
-        paths.append(
-            dict([(f'{input_key}_path', input_path),
-                  (f'{gt_key}_path', gt_path)]))
-    return paths
