+++ ppgan/models/generators/stylegan2_clean_arch.py
@@ -1,10 +1,22 @@
+#   Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserve.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import math
 import random
+
 import paddle
-# from basicsr.archs.arch_util import default_init_weights
 from paddle import nn
 from paddle.nn import functional as F
-import paddle
 
 
 class NormStyleCode(nn.Layer):
