+++ test_tipc/configs/GFPGAN/train_infer_python.txt
@@ -8,12 +8,12 @@ total_iters:lite_train_lite_infer=10
 output_dir:./output/
 dataset.train.batch_size:lite_train_lite_infer=3
 pretrained_model:null
-train_model_name:gfpgan_1024_ffhq*/*checkpoint.pdparams
+train_model_name:gfpgan_ffhq1024*/*checkpoint.pdparams
 train_infer_img_dir:null
 null:null
 ##
 trainer:norm_train
-norm_train:tools/main.py -c configs/gfpgan_1024_ffhq.yaml --seed 123 -o log_config.interval=1 snapshot_config.interval=10
+norm_train:tools/main.py -c configs/gfpgan_ffhq1024.yaml --seed 123 -o log_config.interval=1 snapshot_config.interval=10
 pact_train:null
 fpgm_train:null
 distill_train:null
@@ -27,7 +27,7 @@ null:null
 ===========================infer_params===========================
 --output_dir:./output/
 load:null
-norm_export:tools/export_model.py -c configs/gfpgan_1024_ffhq.yaml  --inputs_size="1,3,512,512"  --model_name inference --load
+norm_export:tools/export_model.py -c configs/gfpgan_ffhq1024.yaml  --inputs_size="1,3,512,512"  --model_name inference --load
 quant_export:null
 fpgm_export:null
 distill_export:null
@@ -37,7 +37,7 @@ inference_dir:inference
 train_model:./inference/stylegan2/stylegan2model_gen
 infer_export:null
 infer_quant:False
-inference:tools/inference.py --model_type gfpgan --seed 123 -c configs/gfpgan_1024_ffhq.yaml --output_path test_tipc/output/ -o validate=None
+inference:tools/inference.py --model_type gfpgan --seed 123 -c configs/gfpgan_ffhq1024.yaml --output_path test_tipc/output/ -o validate=None
 --device:gpu
 null:null
 null:null
